+++ drodrpg/DROD/VerminEffect.cpp
@@ -103,8 +103,8 @@ bool CVerminEffect::Draw(SDL_Surface* pDestSurface)
 			this->dirtyRects[wIndex].w = this->dirtyRects[wIndex].h = 0;
 			continue;
 		}
-		v.fX += cos(v.fAngle) * fMultiplier;
-		v.fY += sin(v.fAngle) * fMultiplier;
+		v.fX += static_cast<float>(cos(v.fAngle)) * fMultiplier;
+		v.fY += static_cast<float>(sin(v.fAngle)) * fMultiplier;
 
 		if (OutOfBounds(v) || HitsObstacle(pRoom, v))
 		{
